delete supplier / show all orders

+++ web/src/components/SupplierCard/SupplierCard.jsx
@@ -1,19 +1,31 @@
+// SupplierCard.js
 import "./supplierCard.css";
 import edit_icon_2 from "../../images/edit-icon-2.png";
 import delete_icon from "../../images/delete-icon.png";
-import { useState, useContext, useEffect } from "react";
+import { useState, useContext } from "react";
 import ConfirmationModal from "../ConfirmationModal/ConfirmationModal";
 import { Context } from "../../uttils/FetchContextProvider";
-import { useParams } from "react-router-dom";
 
 const SupplierCard = ({ supplier }) => {
-  const { id } = useParams();
   const [openConfModal, setOpenConfModal] = useState(false);
-  const { suppliers, setSuppliers } = useContext(Context);
+  const { setSuppliers } = useContext(Context);
+  const [selectedSupplier, setSelectedSupplier] = useState(null);
+
+  const handleDeleteClick = () => {
+    console.log("Delete button clicked"); 
+
+    setSelectedSupplier(supplier._id); 
+    setOpenConfModal(true); 
+  };
 
   const handleDeleteSupplier = async () => {
+    if (!selectedSupplier) {
+      console.log("Supplier not found.//", selectedSupplier);
+      return;
+    }
+    
     try {
-      const response = await fetch(`http://127.0.0.1:10001/suppliers/${id}`, {
+      const response = await fetch(`http://127.0.0.1:10001/suppliers/${selectedSupplier}`, {
         method: "DELETE",
         headers: {
           Authorization: `Bearer ${localStorage.getItem("token")}`,
@@ -21,23 +33,17 @@ const SupplierCard = ({ supplier }) => {
       });
 
       if (response.ok) {
-        setSuppliers((prevSuppliers) =>
-          prevSuppliers.filter((supplier) => supplier.id !== id)
+        setSuppliers(prevSuppliers =>
+          prevSuppliers.filter(supplier => supplier._id !== selectedSupplier)
         );
         setOpenConfModal(false);
       } else {
         console.log("Failed to delete supplier");
       }
     } catch (err) {
-      console.log("Error deleting supplier:", error);
+      console.log("Error deleting supplier:", err);
     }
   };
-
-  useEffect(() => {
-    if (openConfModal) {
-      handleDeleteSupplier();
-    }
-  }, [openConfModal]);
 
   return (
     <div className="Supplier-card">
@@ -58,20 +64,21 @@ const SupplierCard = ({ supplier }) => {
         </p>
         <hr className="hrClass" />
         <span className="supplier-buttons">
-          <img src={edit_icon_2} id="edit_icon" />
+          <img src={edit_icon_2} id="edit_icon" alt="Edit Icon" />
           <img
             src={delete_icon}
             id="delete_icon"
-            onClick={() => setOpenConfModal(true)}
+            onClick={handleDeleteClick}
+            alt="Delete Icon"
           />
         </span>
       </div>
       {openConfModal && (
         <ConfirmationModal
-          closeModal={setOpenConfModal}
+          closeModal={() => setOpenConfModal(false)}
           content="Do you want to delete this supplier?"
           buttonName="CONFIRM"
-          handleConfirm={handleDeleteSupplier}
+          handleConfirm={handleDeleteSupplier} 
         />
       )}
     </div>
